requirements.txt
Flask
requests
gunicorn
vastai
python-dotenv
cryptography
